feat: restructure in packages to test signals
